Added two commands to mr, addAddon2settings and addon2svn
From an addon top directory, running:

    mr addAddon2settings

Will add the name of the current addon to all settings files.
It is important to make sure that the addon conforms with the addon naming guidelines before doing this.
If the key already exists in a settings file, it will not be overwritten.

    mr addon2svn

Will create the addon pot files, and then check which languages have requested the addon, and provide it to them.
If they already have the po file for the addon, any new messages will be merged in.