addon2svn, svn2addon: Fetch and reset to origin/stable before doing anything.
Previously, this relied on mr up being run earlier, but the way the cron jobs are set up, that could have been hours before. That meant that someone might have pushed something after we updated. This mitigates that for the most part.
Resetting rather than pulling ensures we always have the latest upstream, even if there was a force push or a previous run failed in an unexpected state.