Fix git checkout failure (#25)
Git has been updated to return an error code when checkout fails.

When there are no local modifications to the addon, the "stash" logic
is not run. This results in $curBranch being empty. Then, when restoring
the branch the comand git checkout "" is run which results in an error
code 128 and message:
"fatal: empty string is not a valid pathspec. please use . instead if you meant to match all paths"