Added one level when searching for .py files to construct the .pot file (#107)
Another issue found on the .pot generation

Issue
Some strings of the add-on store GUI are not provided to translators as reported in this message.

Cause
When generating the .pot file, svn2nvda looks for .py files only in NVDA's source folder and maximum two levels below. However, after merging of nvaccess/nvda#13985, the GUI files of the add-on store are three levels below, thus they are not taken as source when generating the .pot file.

Solution
Added one more level for the sources used by xgettext to generate the .pot file.